Provider: Moved files accordingly
