+++ lib/pages/docs.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:minitel_toolbox/components/drawer.dart';
-import 'package:minitel_toolbox/components/page_animation.dart';
+import 'package:minitel_toolbox/ui/widgets/drawer.dart';
+import 'package:minitel_toolbox/ui/widgets/page_animation.dart';
 import 'package:minitel_toolbox/pages/docs_pages/minitel_doc.dart';
 import 'package:minitel_toolbox/pages/docs_pages/toolbox_docs.dart';
 import 'package:minitel_toolbox/pages/docs_pages/wiki_docs.dart';
