+++ lib/pages/docs_pages/toolbox_docs/diagnose_doc.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:minitel_toolbox/components/cards.dart';
+import 'package:minitel_toolbox/ui/widgets/cards.dart';
 
 class DiagnoseDoc extends StatelessWidget {
   const DiagnoseDoc({
