+++ lib/pages/maps.dartx
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:minitel_toolbox/components/drawer.dart';
+import 'package:minitel_toolbox/ui/widgets/drawer.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 class MapsPage extends StatefulWidget {
