+++ lib/ui/shared/app_colors.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+const Color primaryColor = Colors.green;
+const Color fontColor = Colors.black87;
+const Color terminalHeaderColor = Colors.lightBlue;
+const Color terminalBgColor = Colors.black87;
+const Color terminalFgColor = Colors.white;
+const Color reportPrimaryColor = Colors.red;
